Moved general asserts to `gen` package
