+++ aarch64/isa/bcond.go
@@ -23,7 +23,7 @@ func NewBcond(condition immediates.Condition) gen.InstructionDefinition {
 func (b Bcond) Target(
 	info *gen.InstructionInfo,
 ) (*gen.LabelInfo, core.ResultList) {
-	results := aarch64translation.AssertArgumentsExactly(info, 1)
+	results := gen.AssertArgumentsExactly(info, 1)
 	if !results.IsEmpty() {
 		return nil, results
 	}
@@ -91,7 +91,7 @@ func (b Bcond) Validate(
 	_, curResults := b.Target(info)
 	results.Extend(&curResults)
 
-	curResults = aarch64translation.AssertTargetsExactly(info, 0)
+	curResults = gen.AssertTargetsExactly(info, 0)
 	results.Extend(&curResults)
 
 	if !results.IsEmpty() {
